add make-based build system
A commit introducing a makefile in a go project may be somewhat
controversial, so here's my position:
Building a project is not only compiling; there may be some other tasks
to carry out, such as building documentation, running some analysis tool
or even building subprojects and so on - as far as I understand (which
doesn't imply I'm not wrong) 'go build' doesn't suffice in those
situations. make or some equivalent build system is always (this is an
intentionally strong assertion) needed for moderately complex projects.

In this particular case, being cross-compilation non-trivial, the
provided makefile also documents the options needed to complete the
task.

Future work should focus on properly listing source files so that we can
avoid useless builds when nothing has changed.

Signed-off-by: Davide "FunkyAss" Del Zompo <davide.delzompo@gmail.com>